Move other migration tests to respective files
