+++ tests/routers/openml/task_list_test.py
@@ -1,12 +1,15 @@
+import asyncio
 from http import HTTPStatus
-from typing import Any
+from typing import Any, cast
 
+import deepdiff
 import httpx
 import pytest
 from sqlalchemy.ext.asyncio import AsyncConnection
 
+from core.conversions import nested_remove_single_element_list
 from core.errors import NoResultsError
-from routers.dependencies import Pagination
+from routers.dependencies import LIMIT_MAX, Pagination
 from routers.openml.tasks import TaskStatusFilter, list_tasks
 
 
@@ -250,3 +253,162 @@ async def test_list_tasks_no_results(payload: dict[str, Any], expdb_test: AsyncC
     """Filters matching nothing return 404 NoResultsError."""
     with pytest.raises(NoResultsError):
         await list_tasks(pagination=Pagination(), expdb=expdb_test, **payload)
+
+
+_TASK_LIST_NO_RESULTS_CODE = "482"
+
+
+def _build_php_task_list_path(php_params: dict[str, Any]) -> str:
+    """Build a PHP-style path for /task/list with path-encoded filter parameters."""
+    if not php_params:
+        return "/task/list"
+    parts = "/".join(f"{k}/{v}" for k, v in php_params.items())
+    return f"/task/list/{parts}"
+
+
+def _normalize_py_task(task: dict[str, Any]) -> dict[str, Any]:
+    """Normalize a single Python task list entry to match PHP format.
+
+    PHP (XML-to-JSON) returns single-element arrays as plain values, not lists.
+    PHP returns task_id, task_type_id, and did as integers (same for Python).
+    and completely omits the "tag" field for all tasks in the list endpoint.
+    """
+    t = nested_remove_single_element_list(task.copy())
+
+    # PHP's list endpoint does not return tags AT ALL
+    t.pop("tag", None)
+
+    # PHP omits qualities where value is None string
+    if "quality" in t:
+        t["quality"] = [q for q in t["quality"] if q.get("value") != "None"]
+
+    return cast("dict[str, Any]", t)
+
+
+_FILTER_COMBOS: list[tuple[dict[str, Any], dict[str, Any]]] = [
+    ({"type": 1}, {"task_type_id": 1}),  # by task type
+    ({"tag": "OpenML100"}, {"tag": "OpenML100"}),  # by tag
+    ({"type": 1, "tag": "OpenML100"}, {"task_type_id": 1, "tag": "OpenML100"}),  # combined
+    ({"data_name": "iris"}, {"data_name": "iris"}),  # by dataset name
+    ({"data_id": 61}, {"data_id": [61]}),  # by dataset id
+    ({"data_tag": "study_14"}, {"data_tag": "study_14"}),  # by dataset tag
+    ({"number_instances": "150"}, {"number_instances": "150"}),  # quality filter
+    (
+        {"data_id": 61, "number_instances": "150"},
+        {"data_id": [61], "number_instances": "150"},
+    ),
+]
+_FILTER_IDS = [
+    "type",
+    "tag",
+    "type_and_tag",
+    "data_name",
+    "data_id",
+    "data_tag",
+    "number_instances",
+    "data_and_quality",
+]
+
+
+@pytest.mark.parametrize(
+    ("php_params", "py_extra"),
+    _FILTER_COMBOS,
+    ids=_FILTER_IDS,
+)
+async def test_list_tasks_equal(
+    php_params: dict[str, Any],
+    py_extra: dict[str, Any],
+    py_api: httpx.AsyncClient,
+    php_api: httpx.AsyncClient,
+) -> None:
+    """Python and PHP task list responses contain the same tasks for the same filters.
+
+    Known differences documented here:
+    - PHP wraps response in {"tasks": {"task": [...]}}, Python returns a flat list.
+    - PHP uses XML-to-JSON which collapses single-element arrays into plain values.
+    - PHP omits the "tag" key when a task has no tags; Python returns "tag": [].
+    - PHP error status is 412 PRECONDITION_FAILED; Python uses 404 NOT_FOUND.
+    """
+    php_path = _build_php_task_list_path(php_params)
+    py_body = {**py_extra, "pagination": {"limit": LIMIT_MAX, "offset": 0}}
+    py_response, php_response = await asyncio.gather(
+        py_api.post("/tasks/list", json=py_body),
+        php_api.get(php_path),
+    )
+
+    # Error case: no results — PHP returns 412, Python returns 404
+    if php_response.status_code == HTTPStatus.PRECONDITION_FAILED:
+        assert py_response.status_code == HTTPStatus.NOT_FOUND
+        assert py_response.headers["content-type"] == "application/problem+json"
+        assert php_response.json()["error"]["code"] == _TASK_LIST_NO_RESULTS_CODE
+        assert py_response.json()["code"] == _TASK_LIST_NO_RESULTS_CODE
+        return
+
+    assert php_response.status_code == HTTPStatus.OK
+    assert py_response.status_code == HTTPStatus.OK
+
+    php_tasks_raw = php_response.json()["tasks"]["task"]
+    php_tasks: list[dict[str, Any]] = (
+        php_tasks_raw if isinstance(php_tasks_raw, list) else [php_tasks_raw]
+    )
+    php_tasks = php_tasks[:LIMIT_MAX]
+    py_tasks: list[dict[str, Any]] = [_normalize_py_task(t) for t in py_response.json()]
+
+    php_ids = {int(t["task_id"]) for t in php_tasks}
+    py_ids = {int(t["task_id"]) for t in py_tasks}
+
+    assert py_ids == php_ids, (
+        f"PHP and Python must return the exact same task IDs: {php_ids ^ py_ids}"
+    )
+
+    # Compare only the tasks PHP returned — per-task deepdiff for clear error messages
+    py_by_id = {int(t["task_id"]): t for t in py_tasks}
+    php_by_id = {int(t["task_id"]): t for t in php_tasks}
+    for task_id in php_ids:
+        differences = deepdiff.diff.DeepDiff(
+            py_by_id[task_id],
+            php_by_id[task_id],
+            ignore_order=True,
+        )
+        assert not differences, f"Differences for task {task_id}: {differences}"
+
+
+@pytest.mark.parametrize(
+    ("php_params", "py_extra"),
+    [
+        ({"tag": "nonexistent_tag_xyz_abc"}, {"tag": "nonexistent_tag_xyz_abc"}),
+        ({"type": 9999}, {"task_type_id": 9999}),
+        ({"data_name": "nonexistent_dataset_xyz"}, {"data_name": "nonexistent_dataset_xyz"}),
+    ],
+    ids=["bad_tag", "bad_type", "bad_data_name"],
+)
+async def test_list_tasks_no_results_matches_php(
+    php_params: dict[str, Any],
+    py_extra: dict[str, Any],
+    py_api: httpx.AsyncClient,
+    php_api: httpx.AsyncClient,
+) -> None:
+    """Both APIs return a "no results" error for filters matching nothing.
+
+    Documented differences:
+    - PHP returns 412 PRECONDITION_FAILED; Python returns 404 NOT_FOUND.
+    - PHP message: "No results"; Python detail: "No tasks match the search criteria."
+    """
+    php_path = _build_php_task_list_path(php_params)
+    py_response, php_response = await asyncio.gather(
+        py_api.post("/tasks/list", json=py_extra),
+        php_api.get(php_path),
+    )
+
+    assert php_response.status_code == HTTPStatus.PRECONDITION_FAILED
+    assert py_response.status_code == HTTPStatus.NOT_FOUND
+
+    php_error = php_response.json()["error"]
+    py_error = py_response.json()
+
+    # Error codes should be the same
+    assert php_error["code"] == _TASK_LIST_NO_RESULTS_CODE
+    assert py_error["code"] == _TASK_LIST_NO_RESULTS_CODE
+    assert php_error["message"] == "No results"
+    assert py_error["detail"] == "No tasks match the search criteria."
+    assert py_response.headers["content-type"] == "application/problem+json"
